fixed view menu error when manager creates new resaurant id adds item and tries to view
