+++ Backend/FastAPI_DB/data/menus.json
@@ -15079,5 +15079,9 @@
         "price": 24.49
       }
     ]
+  },
+  {
+    "menu_id": 102,
+    "items": []
   }
 ]
