+++ Backend/FastAPI_DB/data/users.json
@@ -25,7 +25,7 @@
     "username": "aiden",
     "password": "123",
     "type": 3,
-    "restaurantId": 101
+    "restaurantId": 102
   },
   {
     "id": "b84434dc-84e3-4d0f-b80c-6e233ab1b5e7",
