+++ Backend/User/user_utils.py
@@ -205,84 +205,6 @@ def alter_order_json(new_order: Order):
 			break
 	save_orders(orders)
 		
-def view_menu(restaurant_id):
-	menu = get_menu_by_menu_ID(str(restaurant_id)).items
-	empty = 0
-	if len(menu) == empty:
-		print("no items/ no menu")
-		return
-	print("Item id \t\t Item name \t\t Price")
-	for item in menu:
-		print(str(item.item_id) + " \t\t" + item.name + " \t\t $" + str(item.price)) 
-
-def add_menu_item(restaurant_id): 
-	"""
-	Adds a new item created by the manager to the menu and item database
-
-	checks that item name is valid
-		and that a valid float and positive price check
-	"""
-	print("Enter new item name or 'q' to exit")
-	while(True):
-		item_name = input()
-		if item_name == 'q':
-			return
-		elif(item_name.isalpha() or (len(item_name)>4 and len(item_name)<20)): 
-			break
-		print("Invalid entry try again:")
-
-	print("Enter new item price or 'q' to exit")
-	while(True):
-		item_price = input()
-		if item_price == 'q':
-			return 
-		elif(isFloat(item_price) and float(item_price) > 0): 
-			break
-		print("Invalid entry try again:")
-
-	item_id = str(restaurant_id) + "-" + item_name
-	new_Item = ItemCreate(item_id = item_id, restaurant_id = restaurant_id, name = item_name, tags = [], price = float(item_price))
-	create_items(new_Item)
-
-	menus = load_menu()
-	for menu in menus:
-		if menu["menu_id"] == restaurant_id:
-			menu["items"].append(new_Item.model_dump()) #note to self:model dump converts the model to dict
-	save_menus(menus)
-	print("Item added to menu and item database!")
-
-def remove_menu_item(restaurant_id):
-	"""
-	Removes an item from the menu
-	"""
-	menu = get_menu_by_menu_ID(str(restaurant_id)).items
-	if len(menu) == 0:
-		print("No items/ no menu")
-		return
-	print("Enter item id to remove or 'q' to exit")
-	while(True):
-		user_input = input()
-		if user_input == 'q':
-			return
-		else:
-			match = False 
-			for item in menu:
-				if item.item_id == user_input:
-					match = True
-					break
-			if match:
-				menus = load_menu()
-				for menu in menus:
-					if menu["menu_id"] == restaurant_id:
-						menu["items"] = [item for item in menu["items"] if item["item_id"] != user_input]
-						break
-				save_menus(menus)
-				items = load_items()
-				items = [item for item in items if item["item_id"] != user_input] 
-				save_items(items)
-				print("Item removed from menu and item database")
-				return
-			print("No such item id in menu, try again:")
 
 def isFloat(string): 
 	"""
